refactor: Remove legacy plugin system and clean up deprecated code
- Delete deprecated-types.ts, PromptBasedRouter.ts, RoutingPlan.ts
- Replace BasePlugin.ts with SimpleBasePlugin.ts (clean implementation)
- Make PipelinePlugin standalone (no longer inherits from BasePlugin)
- Remove unused types: ChunkFinding, GlobalFinding, DocumentProfile, RoutingDecision
- Remove unused methods: getTextBefore/getTextAfter from TextChunk
- Remove unused functions: convertLocatedToComments, createPluginError, createFindingsSummary
- Delete example-usage.ts (never imported anywhere)
- Fix LLM interaction tracking in all plugins (Math, Spelling, FactCheck, Forecast)
- Update plugins to manually track interactions and costs from extractWithTool
- Remove LocationTracking mixin (no longer needed)
- All tests passing after refactoring
