+++ src/lib/documentAnalysis/plugin-system/plugins/math/index.ts
@@ -89,7 +89,12 @@ export class MathPlugin extends PipelinePlugin<MathFindingStorage> {
     });
     
     // Track the interaction and cost using parent method
-    await this.trackLLMCall(async () => extraction);
+    if (extraction.interaction) {
+      this.analysisInteractions.push(extraction.interaction);
+    }
+    if (extraction.cost) {
+      this.totalCost += extraction.cost;
+    }
 
     const newFindings = MathHelpers.convertMathResults(
       extraction.result.items || [],
