+++ src/lib/documentAnalysis/plugin-system/plugins/spelling/index.ts
@@ -90,7 +90,12 @@ export class SpellingPlugin extends PipelinePlugin<SpellingFindingStorage> {
     });
     
     // Track the interaction and cost using parent method
-    await this.trackLLMCall(async () => extraction);
+    if (extraction.interaction) {
+      this.analysisInteractions.push(extraction.interaction);
+    }
+    if (extraction.cost) {
+      this.totalCost += extraction.cost;
+    }
 
     const newFindings = this.convertToFindings(
       extraction.result.errors || [],
