+++ src/lib/documentAnalysis/plugin-system/types.ts
@@ -21,8 +21,6 @@ export interface TextChunk {
   
   // Helper methods
   getContext(position: number, windowSize?: number): string;
-  getTextBefore(length: number): string;
-  getTextAfter(length: number): string;
   getLineNumber(charOffset: number): number | null; // Get line number for a character offset within the chunk
 }
 
@@ -55,8 +53,6 @@ export interface LocatedFinding extends Omit<Finding, 'locationHint'> {
   };
 }
 
-export type ChunkFinding = LocatedFinding;
-
 // New finding system types
 export interface HighlightHint {
   searchText: string;      // The text to search for
@@ -85,14 +81,6 @@ export interface InvestigatedFinding {
   highlightHint: HighlightHint;
 }
 
-export interface GlobalFinding {
-  id: string;
-  type: string;
-  data: FindingData;
-  severity: 'low' | 'medium' | 'high' | 'info';
-  message: string;
-  reason: string;  // Why this is global (e.g., "Cross-document pattern")
-}
 
 // Note: PluginError with legacy phase types has been removed (see BasePlugin.ts for legacy support)
 // For new error tracking, use the modern phase types: 'scanChunk' | 'investigate' | 'synthesize' | 'comment'
@@ -137,14 +125,4 @@ export interface SimpleAnalysisPlugin {
   getDebugInfo?(): Record<string, unknown>;
   getCost(): number;
   getLLMInteractions(): LLMInteraction[];
-}
-
-export interface DocumentProfile {
-  documentType?: string;
-  topics?: string[];
-  language?: string;
-  sections?: string[];
-  hasFormulas?: boolean;
-  hasCitations?: boolean;
-  estimatedComplexity?: 'low' | 'medium' | 'high';
 }
