+++ src/lib/documentAnalysis/utils/LocationUtils.ts
@@ -3,6 +3,9 @@
  * Used by both plugins and highlight extraction.
  */
 
+import { TextChunk } from '../plugin-system/TextChunk';
+import { Finding } from '../plugin-system/types';
+
 export interface LineInfo {
   lineNumber: number; // 1-based line number for human readability
   lineIndex: number; // 0-based line index for array access
@@ -14,6 +17,17 @@ export interface LocationInfo {
   start: LineInfo;
   end: LineInfo;
   text: string; // The highlighted text
+}
+
+export interface LocationInfoExtended {
+  lineNumber?: number;
+  lineText?: string;
+  startLine?: number;
+  endLine?: number;
+  position?: {
+    start: number;
+    end: number;
+  };
 }
 
 export class LocationUtils {
@@ -172,5 +186,162 @@ export class LocationUtils {
     }
 
     return null;
+  }
+
+  /**
+   * Find location information for a text match within a chunk
+   */
+  static findLocation(
+    text: string,
+    chunk: TextChunk,
+    startSearchFrom: number = 0
+  ): LocationInfoExtended | null {
+    const chunkLocationUtils = new LocationUtils(chunk.text);
+    const textPosition = chunk.text.indexOf(text, startSearchFrom);
+    
+    if (textPosition === -1) {
+      return null;
+    }
+
+    const locationInfo = chunkLocationUtils.getLocationInfo(
+      textPosition,
+      textPosition + text.length
+    );
+
+    if (!locationInfo) {
+      return null;
+    }
+
+    const result: LocationInfoExtended = {
+      position: {
+        start: textPosition,
+        end: textPosition + text.length
+      }
+    };
+
+    // Calculate absolute line numbers if chunk has global line info
+    if (chunk.metadata?.lineInfo) {
+      result.lineNumber = chunk.metadata.lineInfo.startLine + locationInfo.start.lineNumber - 1;
+      result.startLine = result.lineNumber;
+      result.endLine = chunk.metadata.lineInfo.startLine + locationInfo.end.lineNumber - 1;
+    } else {
+      // Use chunk-relative line numbers
+      result.lineNumber = locationInfo.start.lineNumber;
+      result.startLine = locationInfo.start.lineNumber;
+      result.endLine = locationInfo.end.lineNumber;
+    }
+
+    result.lineText = locationInfo.start.lineText;
+
+    return result;
+  }
+
+  /**
+   * Add location hint to a finding
+   */
+  static addLocationHint(
+    finding: Finding,
+    text: string,
+    chunk: TextChunk,
+    startSearchFrom: number = 0
+  ): Finding {
+    const location = this.findLocation(text, chunk, startSearchFrom);
+    
+    if (location && location.lineNumber && location.lineText) {
+      finding.locationHint = {
+        lineNumber: location.lineNumber,
+        lineText: location.lineText,
+        matchText: text,
+        startLineNumber: location.startLine,
+        endLineNumber: location.endLine
+      };
+    }
+
+    if (location?.position) {
+      finding.location = location.position;
+    }
+
+    return finding;
+  }
+
+  /**
+   * Create a finding with automatic location tracking
+   */
+  static createFindingWithLocation(
+    type: string,
+    severity: 'low' | 'medium' | 'high' | 'info',
+    message: string,
+    matchText: string,
+    chunk: TextChunk,
+    metadata?: Record<string, any>
+  ): Finding {
+    const finding: Finding = {
+      type,
+      severity,
+      message,
+      metadata: {
+        ...metadata,
+        chunkId: chunk.id
+      }
+    };
+
+    return this.addLocationHint(finding, matchText, chunk);
+  }
+
+  /**
+   * Find all occurrences of a text in a chunk
+   */
+  static findAllLocations(
+    text: string,
+    chunk: TextChunk
+  ): LocationInfoExtended[] {
+    const locations: LocationInfoExtended[] = [];
+    let searchFrom = 0;
+    
+    while (true) {
+      const location = this.findLocation(text, chunk, searchFrom);
+      if (!location || !location.position) break;
+      
+      locations.push(location);
+      searchFrom = location.position.end;
+    }
+    
+    return locations;
+  }
+
+  /**
+   * Get location info for a specific character range in a chunk
+   */
+  static getLocationForRange(
+    start: number,
+    end: number,
+    chunk: TextChunk
+  ): LocationInfoExtended | null {
+    const chunkLocationUtils = new LocationUtils(chunk.text);
+    const locationInfo = chunkLocationUtils.getLocationInfo(start, end);
+
+    if (!locationInfo) {
+      return null;
+    }
+
+    const result: LocationInfoExtended = {
+      position: { start, end }
+    };
+
+    // Calculate absolute line numbers if chunk has global line info
+    if (chunk.metadata?.lineInfo) {
+      result.lineNumber = chunk.metadata.lineInfo.startLine + locationInfo.start.lineNumber - 1;
+      result.startLine = result.lineNumber;
+      result.endLine = chunk.metadata.lineInfo.startLine + locationInfo.end.lineNumber - 1;
+    } else {
+      // Use chunk-relative line numbers
+      result.lineNumber = locationInfo.start.lineNumber;
+      result.startLine = locationInfo.start.lineNumber;
+      result.endLine = locationInfo.end.lineNumber;
+    }
+
+    result.lineText = locationInfo.start.lineText;
+
+    return result;
   }
 }
